pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "spawn-agent"
version = "0.4.2"
description = "Real-time on-chain intelligence agent with OTP-inspired supervision trees"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "SpawnAgent Contributors"},
]
keywords = [
    "blockchain",
    "monitoring",
    "on-chain",
    "wallet-tracking",
    "defi",
    "analytics",
    "real-time",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: System :: Monitoring",
    "Typing :: Typed",
    "Framework :: AsyncIO",
]

dependencies = [
    "aiohttp>=3.9,<4.0",
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "black>=24.0",
    "ruff>=0.3",
    "mypy>=1.8",
    "pre-commit>=3.6",
]
alerts = [
    "aiohttp>=3.9",
]

[project.scripts]
spawn-agent = "spawn_agent.cli.commands:cli"

[project.urls]
Homepage = "https://github.com/Hstkj23/GigaVault"
Documentation = "https://github.com/Hstkj23/GigaVault#readme"
Repository = "https://github.com/Hstkj23/GigaVault"
Issues = "https://github.com/Hstkj23/GigaVault/issues"

[tool.setuptools.packages.find]
include = ["spawn_agent*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests requiring RPC access",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "RUF", # ruff-specific rules
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["spawn_agent"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
